requirements.txt
streamlit
roboflow
ultralytics
opencv-python-headless
pillow
numpy
python-dotenv
